Refactor: package structure for dataverse featured items
